refactor: split KV cache utilities from tasks/_base.py into tasks/_cache_utils.py
Move ~350 lines of KV cache and linear attention utilities out of _base.py
into a new focused module _cache_utils.py:

- StatePair type alias (tuple[ir.Value, ...] | tuple[None, None])
- LinearAttentionDims NamedTuple + linear_attention_dims()
- _make_kv_cache_inputs() / _register_kv_cache_outputs()
- _make_hybrid_cache_inputs() / _register_hybrid_cache_outputs()
  (includes conv, lightning_attention, linear_attention, mamba, mamba2)
- _register_linear_attention_functions()
- _FUNCTIONS_DOMAIN constant

_base.py retains ModelTask ABC (with model_roles ClassVar), _make_graph(),
and _make_model(). Caller files (_causal_lm, _vision_language, _multimodal,
_phi4mm_multimodal, _speech_language, _seq2seq, _tts, _vision_language_3model,
_speech_to_text) updated to import cache symbols directly from _cache_utils.

2293 tests pass.
