+++ src/mobius/tasks/_base.py
@@ -6,7 +6,7 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import ClassVar, NamedTuple
+from typing import ClassVar
 
 import onnx_ir as ir
 from onnxscript import nn
@@ -16,51 +16,6 @@ import mobius
 from mobius._configs import BaseModelConfig
 from mobius._constants import OPSET_VERSION
 from mobius._model_package import ModelPackage
-
-_FUNCTIONS_DOMAIN = "pkg.mobius"
-
-# Cache state pair: (key, value) or (conv_state, ssm_state) for stateful
-# layers.  Single-state layers (conv/lightning) use a 1-tuple.
-# MLP-only layers are stateless and use (None, None).
-StatePair = tuple[ir.Value, ir.Value] | tuple[ir.Value] | tuple[None, None]
-
-
-class LinearAttentionDims(NamedTuple):
-    """Dimension sizes for linear attention (DeltaNet) layers."""
-
-    num_k_heads: int
-    num_v_heads: int
-    head_k_dim: int
-    head_v_dim: int
-    key_dim: int  # = head_k_dim * num_k_heads
-    value_dim: int  # = head_v_dim * num_v_heads
-    conv_dim: int  # = key_dim * 2 + value_dim
-    conv_kernel: int
-
-
-def linear_attention_dims(config: BaseModelConfig) -> LinearAttentionDims:
-    """Compute dimension sizes for linear attention from config.
-
-    Raises ``TypeError`` if any required config field is ``None``.
-    """
-    num_k_heads = config.linear_num_key_heads
-    num_v_heads = config.linear_num_value_heads
-    head_k_dim = config.linear_key_head_dim
-    head_v_dim = config.linear_value_head_dim
-    conv_kernel = config.linear_conv_kernel_dim
-    key_dim = head_k_dim * num_k_heads
-    value_dim = head_v_dim * num_v_heads
-    conv_dim = key_dim * 2 + value_dim
-    return LinearAttentionDims(
-        num_k_heads=num_k_heads,
-        num_v_heads=num_v_heads,
-        head_k_dim=head_k_dim,
-        head_v_dim=head_v_dim,
-        key_dim=key_dim,
-        value_dim=value_dim,
-        conv_dim=conv_dim,
-        conv_kernel=conv_kernel,
-    )
 
 
 def _make_graph(
@@ -88,70 +43,6 @@ def _make_model(graph: ir.Graph) -> ir.Model:
     model.producer_name = "mobius"
     model.producer_version = mobius.__version__
     return model
-
-
-def _make_kv_cache_inputs(
-    num_layers: int,
-    num_kv_heads: int,
-    head_dim: int,
-    dtype: ir.DataType,
-    batch: ir.SymbolicDim,
-    past_seq_len: ir.SymbolicDim,
-    *,
-    prefix: str = "past_key_values",
-    key_head_dim: int | None = None,
-    value_head_dim: int | None = None,
-) -> tuple[list[ir.Value], list[tuple[ir.Value, ir.Value]]]:
-    """Create KV cache input values for ``num_layers`` layers.
-
-    Args:
-        key_head_dim: Head dim for keys. Defaults to ``head_dim``.
-            For MLA attention, this is ``qk_nope_head_dim + qk_rope_head_dim``.
-        value_head_dim: Head dim for values. Defaults to ``head_dim``.
-            For MLA attention, this is ``v_head_dim``.
-
-    Returns:
-        ``(flat_inputs, kv_pairs)`` where *flat_inputs* is a flat list
-        suitable for extending ``graph_inputs`` and *kv_pairs* is a list
-        of ``(key, value)`` tuples for passing to the module.
-    """
-    k_dim = key_head_dim if key_head_dim is not None else head_dim
-    v_dim = value_head_dim if value_head_dim is not None else head_dim
-    flat: list[ir.Value] = []
-    pairs: list[tuple[ir.Value, ir.Value]] = []
-    for i in range(num_layers):
-        past_key = ir.Value(
-            name=f"{prefix}.{i}.key",
-            shape=ir.Shape([batch, num_kv_heads, past_seq_len, k_dim]),
-            type=ir.TensorType(dtype),
-        )
-        past_value = ir.Value(
-            name=f"{prefix}.{i}.value",
-            shape=ir.Shape([batch, num_kv_heads, past_seq_len, v_dim]),
-            type=ir.TensorType(dtype),
-        )
-        flat.extend([past_key, past_value])
-        pairs.append((past_key, past_value))
-    return flat, pairs
-
-
-def _register_kv_cache_outputs(
-    graph: ir.Graph,
-    present_key_values: list[tuple[ir.Value, ir.Value]],
-    *,
-    prefix: str = "present",
-) -> None:
-    """Name and register KV cache outputs on the graph.
-
-    Output shapes and dtypes are inferred by the shape inference pass
-    that runs during model optimization.
-    """
-    for i, (present_key, present_value) in enumerate(present_key_values):
-        present_key.name = f"{prefix}.{i}.key"
-        present_value.name = f"{prefix}.{i}.value"
-
-        graph.outputs.append(present_key)
-        graph.outputs.append(present_value)
 
 
 class ModelTask(ABC):
@@ -187,246 +78,3 @@ class ModelTask(ABC):
             A :class:`ModelPackage` containing the built model(s).
         """
         ...
-
-
-def _make_hybrid_cache_inputs(
-    config: BaseModelConfig,
-    dtype: ir.DataType,
-    batch: ir.SymbolicDim,
-    past_seq_len: ir.SymbolicDim,
-    *,
-    prefix: str = "past_key_values",
-) -> tuple[list[ir.Value], list[StatePair]]:
-    """Create cache inputs for hybrid models with mixed layer types.
-
-    Supported layer types:
-        ``"full_attention"`` — standard KV cache (key + value).
-        ``"conv"`` — ShortConv conv_state only.
-        ``"linear_attention"`` (DeltaNet) — conv_state + recurrent_state.
-        ``"mamba"`` / ``"mamba2"`` — conv_state + ssm_state.
-        ``"mlp"`` — stateless, produces ``(None, None)`` pair.
-
-    Returns:
-        ``(flat_inputs, state_pairs)`` — *flat_inputs* contains only
-        the ``ir.Value`` entries (no graph inputs for MLP layers);
-        *state_pairs* has one entry per layer, with ``(None, None)``
-        for stateless MLP layers.
-    """
-    layer_types = config.layer_types or []
-    flat: list[ir.Value] = []
-    pairs: list[StatePair] = []
-
-    # DeltaNet dimensions from config (computed once via shared helper)
-    has_linear = "linear_attention" in layer_types
-    if has_linear:
-        dims = linear_attention_dims(config)
-
-    # Mamba SSM dimensions from config (Jamba-style)
-    mamba_expand = getattr(config, "mamba_expand", 2)
-    mamba_d_inner = config.hidden_size * mamba_expand
-    mamba_d_conv = getattr(config, "mamba_d_conv", 4)
-    mamba_d_state = getattr(config, "mamba_d_state", 16)
-
-    # Mamba2/SSD dimensions from config (Bamba-style).
-    # Defaults are 0 so a missing field produces a clear shape error
-    # rather than silently using model-specific values.
-    mamba2_n_heads = getattr(config, "mamba_n_heads", 0)
-    mamba2_d_head = getattr(config, "mamba_d_head", 0)
-    mamba2_d_state = getattr(config, "mamba_d_state", 0)
-    mamba2_n_groups = getattr(config, "mamba_n_groups", 1)
-    # Prefer n_heads * d_head (NemotronH); fall back to hidden * expand (Bamba)
-    mamba2_d_inner = (
-        mamba2_n_heads * mamba2_d_head
-        if mamba2_n_heads and mamba2_d_head
-        else config.hidden_size * mamba_expand
-    )
-    mamba2_conv_dim = mamba2_d_inner + 2 * mamba2_n_groups * mamba2_d_state
-
-    for i in range(config.num_hidden_layers):
-        ltype = layer_types[i] if i < len(layer_types) else "full_attention"
-
-        if ltype == "lightning_attention":
-            # Lightning Attention: single recurrent state only (no conv_state)
-            # State: (B, num_heads, head_dim, head_dim) — square matrix accumulator
-            rec_state = ir.Value(
-                name=f"{prefix}.{i}.recurrent_state",
-                shape=ir.Shape(
-                    [batch, config.num_attention_heads, config.head_dim, config.head_dim]
-                ),
-                type=ir.TensorType(dtype),
-            )
-            flat.append(rec_state)
-            pairs.append((rec_state,))  # 1-tuple: lightning has no conv_state
-        elif ltype == "linear_attention":
-            conv_state = ir.Value(
-                name=f"{prefix}.{i}.conv_state",
-                shape=ir.Shape([batch, dims.conv_dim, dims.conv_kernel - 1]),
-                type=ir.TensorType(dtype),
-            )
-            rec_state = ir.Value(
-                name=f"{prefix}.{i}.recurrent_state",
-                shape=ir.Shape([batch, dims.num_v_heads, dims.head_k_dim, dims.head_v_dim]),
-                type=ir.TensorType(dtype),
-            )
-            flat.extend([conv_state, rec_state])
-            pairs.append((conv_state, rec_state))
-        elif ltype == "conv":
-            # ShortConv layers: conv_state only (no SSM state)
-            # State: (batch, hidden_size, short_conv_kernel - 1)
-            short_conv_kernel = getattr(config, "short_conv_kernel", 3)
-            conv_state = ir.Value(
-                name=f"{prefix}.{i}.conv_state",
-                shape=ir.Shape([batch, config.hidden_size, short_conv_kernel - 1]),
-                type=ir.TensorType(dtype),
-            )
-            flat.append(conv_state)
-            pairs.append((conv_state,))  # 1-tuple: conv has no second state
-        elif ltype == "mlp":
-            # MLP-only layers are stateless — no cache inputs needed
-            pairs.append((None, None))
-        elif ltype == "mamba":
-            conv_state = ir.Value(
-                name=f"{prefix}.{i}.conv_state",
-                shape=ir.Shape([batch, mamba_d_inner, mamba_d_conv - 1]),
-                type=ir.TensorType(dtype),
-            )
-            ssm_state = ir.Value(
-                name=f"{prefix}.{i}.ssm_state",
-                shape=ir.Shape([batch, mamba_d_inner, mamba_d_state]),
-                type=ir.TensorType(dtype),
-            )
-            flat.extend([conv_state, ssm_state])
-            pairs.append((conv_state, ssm_state))
-        elif ltype == "mamba2":
-            conv_state = ir.Value(
-                name=f"{prefix}.{i}.conv_state",
-                shape=ir.Shape([batch, mamba2_conv_dim, mamba_d_conv - 1]),
-                type=ir.TensorType(dtype),
-            )
-            ssm_state = ir.Value(
-                name=f"{prefix}.{i}.ssm_state",
-                shape=ir.Shape([batch, mamba2_n_heads, mamba2_d_head, mamba2_d_state]),
-                type=ir.TensorType(dtype),
-            )
-            flat.extend([conv_state, ssm_state])
-            pairs.append((conv_state, ssm_state))
-        else:
-            past_key = ir.Value(
-                name=f"{prefix}.{i}.key",
-                shape=ir.Shape(
-                    [batch, config.num_key_value_heads, past_seq_len, config.head_dim]
-                ),
-                type=ir.TensorType(dtype),
-            )
-            past_value = ir.Value(
-                name=f"{prefix}.{i}.value",
-                shape=ir.Shape(
-                    [batch, config.num_key_value_heads, past_seq_len, config.head_dim]
-                ),
-                type=ir.TensorType(dtype),
-            )
-            flat.extend([past_key, past_value])
-            pairs.append((past_key, past_value))
-
-    return flat, pairs
-
-
-def _register_hybrid_cache_outputs(
-    graph: ir.Graph,
-    present_key_values: list[tuple[ir.Value, ...]],
-    layer_types: list[str],
-    *,
-    prefix: str = "present",
-) -> None:
-    """Name and register hybrid cache outputs on the graph.
-
-    Uses ``.key``/``.value`` for full attention layers,
-    ``.recurrent_state`` for lightning attention layers (1-tuple),
-    ``.conv_state``/``.recurrent_state`` for linear attention layers,
-    and ``.conv_state``/``.ssm_state`` for mamba/mamba2 layers.
-
-    Output shapes and dtypes are inferred by the shape inference pass
-    that runs during model optimization.
-    """
-    for i, states in enumerate(present_key_values):
-        ltype = layer_types[i] if i < len(layer_types) else "full_attention"
-        if ltype == "mlp":
-            continue  # MLP layers produce no cache state
-        if ltype == "lightning_attention":
-            # Single recurrent state only (no conv_state for lightning)
-            (state_a,) = states
-            state_a.name = f"{prefix}.{i}.recurrent_state"
-            graph.outputs.append(state_a)
-        elif ltype == "conv":
-            # ShortConv: single conv_state only
-            (state_a,) = states
-            state_a.name = f"{prefix}.{i}.conv_state"
-            graph.outputs.append(state_a)
-        else:
-            state_a, state_b = states
-            if ltype == "linear_attention":
-                state_a.name = f"{prefix}.{i}.conv_state"
-                state_b.name = f"{prefix}.{i}.recurrent_state"
-            elif ltype in ("mamba", "mamba2"):
-                state_a.name = f"{prefix}.{i}.conv_state"
-                state_b.name = f"{prefix}.{i}.ssm_state"
-            else:
-                state_a.name = f"{prefix}.{i}.key"
-                state_b.name = f"{prefix}.{i}.value"
-
-            graph.outputs.append(state_a)
-            graph.outputs.append(state_b)
-
-
-def _register_linear_attention_functions(
-    model: ir.Model,
-    config: BaseModelConfig,
-) -> None:
-    """Register CausalConvWithState and LinearAttention functions.
-
-    Registers functions for DeltaNet (``linear_attention`` layers) and/or
-    Lightning Attention (``lightning_attention`` layers) as needed.
-    Adds the ``pkg.mobius`` opset import to the graph.
-    """
-    layer_types = getattr(config, "layer_types", None) or []
-    has_deltanet = "linear_attention" in layer_types
-    has_lightning = "lightning_attention" in layer_types
-
-    if not has_deltanet and not has_lightning:
-        return
-
-    from mobius.functions import (
-        causal_conv_nd_with_state,
-        linear_attention,
-    )
-
-    if has_deltanet:
-        dims = linear_attention_dims(config)
-        conv_func = causal_conv_nd_with_state(
-            kernel_size=dims.conv_kernel,
-            channels=dims.conv_dim,
-            ndim=1,
-            activation="silu",
-        )
-        attn_func = linear_attention(
-            q_num_heads=dims.num_k_heads,
-            kv_num_heads=dims.num_v_heads,
-            update_rule="gated_delta",
-            scale=1.0 / (dims.head_k_dim**0.5),
-            stash_type=config.dtype,
-        )
-        model.functions[conv_func.identifier()] = conv_func
-        model.functions[attn_func.identifier()] = attn_func
-
-    if has_lightning:
-        head_dim = config.head_dim
-        attn_func_gated = linear_attention(
-            q_num_heads=config.num_attention_heads,
-            kv_num_heads=config.num_attention_heads,
-            update_rule="gated",
-            scale=1.0 / (head_dim**0.5),
-            stash_type=config.dtype,
-        )
-        model.functions[attn_func_gated.identifier()] = attn_func_gated
-
-    model.graph.opset_imports[_FUNCTIONS_DOMAIN] = 1
